make fixture plugin
