+++ tests/cli/clean/conftest.py
@@ -14,17 +14,6 @@ from cg.models.run_devices.illumina_run_directory_data import IlluminaRunDirecto
 from cg.store.models import IlluminaSequencingRun
 from cg.store.store import Store
 from tests.store_helpers import StoreHelpers
-from tests.services.illumina_services.cleaning_services.conftest import (
-    tmp_clean_dir,
-    tmp_sequencing_run_to_clean,
-    tmp_sequencing_run_not_to_clean,
-    tmp_sequencing_run_to_clean_path,
-    tmp_sequencing_run_not_to_clean_path,
-    housekeeper_api_with_illumina_seq_run_to_clean,
-    hk_illumina_sequencing_run_to_clean_bundle,
-    tmp_sample_sheet_clean_illumina_sequencing_run_path,
-    hk_sample_bundle_for_illumina_sequencing_run_to_clean,
-)
 
 
 @pytest.fixture
@@ -202,35 +191,5 @@ def clean_context_microsalt(
     cg_context.meta_apis["analysis_api"] = analysis_api
 
     cg_context.data_delivery.base_path = f"{project_dir}/rsync"
-
-    return cg_context
-
-
-@pytest.fixture(scope="function")
-def clean_illumina_sequencing_runs_context(
-    cg_context: CGConfig,
-    tmp_clean_dir: Path,
-    tmp_sequencing_run_to_clean: IlluminaRunDirectoryData,
-    store_with_illumina_sequencing_data: Store,
-    housekeeper_api_with_illumina_seq_run_to_clean: HousekeeperAPI,
-    selected_novaseq_6000_pre_1_5_kits_sample_ids: list[str],
-) -> CGConfig:
-    cg_context.run_instruments.illumina.sequencing_runs_dir = tmp_clean_dir
-    cg_context.run_instruments.illumina.demultiplexed_runs_dir = tmp_clean_dir
-    cg_context.housekeeper_api_ = housekeeper_api_with_illumina_seq_run_to_clean
-
-    sequencing_run: IlluminaSequencingRun = (
-        store_with_illumina_sequencing_data.get_illumina_sequencing_run_by_device_internal_id(
-            tmp_sequencing_run_to_clean.id
-        )
-    )
-    sequencing_run.has_backup = True
-    sample_metric_to_delete = [
-        metric
-        for metric in sequencing_run.sample_metrics
-        if metric.sample.internal_id == selected_novaseq_6000_pre_1_5_kits_sample_ids[1]
-    ]
-    store_with_illumina_sequencing_data.session.delete(sample_metric_to_delete[0])
-    cg_context.status_db_ = store_with_illumina_sequencing_data
 
     return cg_context
