+++ tests/fixture_plugins/illumina_clean_fixtures/clean_fixtures.py
@@ -5,11 +5,12 @@ import pytest
 
 from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.constants import SequencingFileTag
+from cg.models.cg_config import CGConfig
 from cg.services.illumina_services.cleaning_services.clean_runs_service import (
     IlluminaCleanRunsService,
 )
 from cg.models.run_devices.illumina_run_directory_data import IlluminaRunDirectoryData
-from cg.store.models import IlluminaSequencingRun, Sample
+from cg.store.models import IlluminaSequencingRun
 from cg.store.store import Store
 from tests.store_helpers import StoreHelpers
 
@@ -223,3 +224,33 @@ def hk_sample_bundle_for_sequencing_run_not_to_clean(
         "expires": timestamp_yesterday,
         "files": [],
     }
+
+
+@pytest.fixture(scope="function")
+def clean_illumina_sequencing_runs_context(
+    cg_context: CGConfig,
+    tmp_clean_dir: Path,
+    tmp_sequencing_run_to_clean: IlluminaRunDirectoryData,
+    store_with_illumina_sequencing_data: Store,
+    housekeeper_api_with_illumina_seq_run_to_clean: HousekeeperAPI,
+    selected_novaseq_6000_pre_1_5_kits_sample_ids: list[str],
+) -> CGConfig:
+    cg_context.run_instruments.illumina.sequencing_runs_dir = tmp_clean_dir
+    cg_context.run_instruments.illumina.demultiplexed_runs_dir = tmp_clean_dir
+    cg_context.housekeeper_api_ = housekeeper_api_with_illumina_seq_run_to_clean
+
+    sequencing_run: IlluminaSequencingRun = (
+        store_with_illumina_sequencing_data.get_illumina_sequencing_run_by_device_internal_id(
+            tmp_sequencing_run_to_clean.id
+        )
+    )
+    sequencing_run.has_backup = True
+    sample_metric_to_delete = [
+        metric
+        for metric in sequencing_run.sample_metrics
+        if metric.sample.internal_id == selected_novaseq_6000_pre_1_5_kits_sample_ids[1]
+    ]
+    store_with_illumina_sequencing_data.session.delete(sample_metric_to_delete[0])
+    cg_context.status_db_ = store_with_illumina_sequencing_data
+
+    return cg_context
